refactor: extract super tenant write actions
