+++ supabase/functions/super-tenant-mutate/index.ts
@@ -26,6 +26,7 @@ import {
 } from "../_shared/validation.ts";
 import type { SuperTenantContext } from "./context.ts";
 import {
+  defaultFeatureFlags,
   enrichTenants as enrichTenantRows,
   handleListTenants,
 } from "./actions/tenantQueries.ts";
@@ -35,6 +36,14 @@ import {
   isMissingDistrictsTable,
   normalizeDistrictSlug,
 } from "./actions/districts.ts";
+import {
+  ensureDistrict,
+  handleTenantWriteAction,
+  TENANT_ACCOUNT_CATEGORIES,
+  TENANT_PLAN_CODES,
+  TENANT_STATUSES,
+  upsertTenantPolicy,
+} from "./actions/tenantWrites.ts";
 
 const baseCorsHeaders = {
   "Access-Control-Allow-Headers":
@@ -131,36 +140,7 @@ const isMissingNamedColumn = (
   return isMissingPostgrestColumn(error, column, { allowSchemaCache: true });
 };
 
-const isMissingFeatureFlagsColumn = (error: PgError | null | undefined) =>
-  isMissingNamedColumn(error, "feature_flags");
-
-const isMissingAccountCategoryColumn = (error: PgError | null | undefined) =>
-  isMissingNamedColumn(error, "account_category");
-
-const isMissingPlanCodeColumn = (error: PgError | null | undefined) =>
-  isMissingNamedColumn(error, "plan_code");
-
-const isTenantPolicyAccountCategoryConstraintError = (
-  error: PgError | null | undefined
-) =>
-  !!error &&
-  (error.code === "23514" || error.code === "P0001") &&
-  (error.message ?? "").toLowerCase().includes("tenant_policies_account_category_check");
-
-const isTenantPolicyPlanCodeConstraintError = (error: PgError | null | undefined) =>
-  !!error &&
-  (error.code === "23514" || error.code === "P0001") &&
-  (error.message ?? "").toLowerCase().includes("tenant_policies_plan_code_check");
-
-const normalizeDistrictName = (value: string | null | undefined) => (value ?? "").trim();
 const lower = (value: string | null | undefined) => (value ?? "").toLowerCase();
-const defaultFeatureFlags = () => ({
-  enable_notifications: true,
-  enable_bulk_item_import: true,
-  enable_bulk_student_tools: true,
-  enable_status_tracking: true,
-  enable_barcode_generator: true,
-});
 
 const resolveCorsHeaders = (req: Request) => {
   const origin = req.headers.get("Origin");
@@ -185,164 +165,7 @@ const resolveResetRedirectTo = (req: Request) => {
   return null;
 };
 
-const isValidStatus = (value: unknown): value is "active" | "suspended" | "archived" =>
-  value === "active" || value === "suspended" || value === "archived";
-const TENANT_STATUSES = new Set(["active", "suspended", "archived"] as const);
-const TENANT_ACCOUNT_CATEGORIES = new Set(["organization", "district", "individual"] as const);
-const TENANT_PLAN_CODES = new Set([
-  "core",
-  "growth",
-  "starter",
-  "scale",
-  "enterprise",
-  "individual_yearly",
-  "individual_monthly",
-] as const);
 
-const verifySuperPassword = async (
-  supabaseUrl: string,
-  publishableKey: string,
-  email: string,
-  password: string
-) => {
-  const authClient = createClient(supabaseUrl, publishableKey, {
-    auth: { persistSession: false },
-  });
-  const { error } = await authClient.auth.signInWithPassword({ email, password });
-  return !error;
-};
-
-const ensureDistrict = async (
-  adminClient: SupabaseAdminClient,
-  districtSlug: string,
-  districtName: string
-) => {
-  const slug = normalizeDistrictSlug(districtSlug);
-  const name = normalizeDistrictName(districtName);
-  if (!slug) {
-    return { districtId: null, error: "District slug is required." };
-  }
-  if (!name) {
-    return { districtId: null, error: "District name is required when assigning a district." };
-  }
-
-  const { data: existing, error: existingError } = await adminClient
-    .from("districts")
-    .select("id, name")
-    .eq("slug", slug)
-    .maybeSingle();
-
-  if (existingError) {
-    if (isMissingDistrictsTable(existingError as PgError)) {
-      return {
-        districtId: null,
-        error: "District foundation is not enabled yet. Run the latest database migration.",
-      };
-    }
-    return { districtId: null, error: "Unable to load district." };
-  }
-
-  if (existing?.id) {
-    if (existing.name !== name) {
-      const { error: updateError } = await adminClient
-        .from("districts")
-        .update({ name, updated_at: new Date().toISOString() })
-        .eq("id", existing.id);
-      if (updateError) {
-        return { districtId: null, error: "Unable to update district." };
-      }
-    }
-    return { districtId: existing.id, error: null };
-  }
-
-  const { data: created, error: createError } = await adminClient
-    .from("districts")
-    .insert({
-      name,
-      slug,
-    })
-    .select("id")
-    .single();
-
-  if (createError || !created?.id) {
-    return { districtId: null, error: "Unable to create district." };
-  }
-
-  return { districtId: created.id, error: null };
-};
-
-type TenantPolicyUpsertInput = {
-  tenant_id: string;
-  checkout_due_hours?: number;
-  account_category?: "organization" | "district" | "individual" | null;
-  plan_code?:
-    | "core"
-    | "growth"
-    | "starter"
-    | "scale"
-    | "enterprise"
-    | "individual_yearly"
-    | "individual_monthly"
-    | null;
-  feature_flags?: Record<string, unknown> | null;
-  updated_by: string;
-  updated_at: string;
-};
-
-const upsertTenantPolicy = async (
-  adminClient: SupabaseAdminClient,
-  policy: TenantPolicyUpsertInput
-) => {
-  const buildPayload = (options: {
-    includeAccountCategory: boolean;
-    includePlanCode: boolean;
-    includeFeatureFlags: boolean;
-  }) => ({
-    tenant_id: policy.tenant_id,
-    checkout_due_hours: policy.checkout_due_hours ?? 72,
-    ...(options.includeAccountCategory ? { account_category: policy.account_category ?? null } : {}),
-    ...(options.includePlanCode ? { plan_code: policy.plan_code ?? null } : {}),
-    ...(options.includeFeatureFlags && policy.feature_flags !== undefined
-      ? { feature_flags: policy.feature_flags ?? defaultFeatureFlags() }
-      : {}),
-    updated_by: policy.updated_by,
-    updated_at: policy.updated_at,
-  });
-
-  let includeAccountCategory = true;
-  let includePlanCode = true;
-  let includeFeatureFlags = true;
-
-  for (;;) {
-    const result = await adminClient.from("tenant_policies").upsert(
-      buildPayload({ includeAccountCategory, includePlanCode, includeFeatureFlags }),
-      { onConflict: "tenant_id" }
-    );
-
-    if (!result.error) return result;
-    if (includeFeatureFlags && isMissingFeatureFlagsColumn(result.error as PgError)) {
-      includeFeatureFlags = false;
-      continue;
-    }
-    if (
-      includeAccountCategory &&
-      (isMissingAccountCategoryColumn(result.error as PgError) ||
-        isTenantPolicyAccountCategoryConstraintError(result.error as PgError))
-    ) {
-      includeAccountCategory = false;
-      continue;
-    }
-    if (
-      includePlanCode &&
-      (isMissingPlanCodeColumn(result.error as PgError) ||
-        isTenantPolicyPlanCodeConstraintError(result.error as PgError))
-    ) {
-      includePlanCode = false;
-      continue;
-    }
-    return result;
-  }
-};
 
 serve(async (req) => {
   const { hasOrigin, originAllowed, headers } = resolveCorsHeaders(req);
@@ -561,7 +384,11 @@ serve(async (req) => {
 
       let districtId: string | null = null;
       if (districtSlug && districtName) {
-        const districtResult = await ensureDistrict(adminClient, districtSlug, districtName);
+        const districtResult = await ensureDistrict(
+          actionContext,
+          districtSlug,
+          districtName
+        );
         if (districtResult.error) {
           return jsonResponse(400, { error: districtResult.error });
         }
@@ -705,7 +532,7 @@ serve(async (req) => {
         }
       }
 
-      const tenantPolicyResult = await upsertTenantPolicy(adminClient, {
+      const tenantPolicyResult = await upsertTenantPolicy(actionContext, {
         tenant_id: data.id,
         checkout_due_hours: 72,
         account_category: accountCategory,
@@ -755,154 +582,9 @@ serve(async (req) => {
       });
     }
 
-    if (action === "update_tenant") {
-      const next = payload;
-      const id = requireUuid(next.id);
-      const name = requireText(next.name, { maxLen: 120 });
-      const accessCode = requireText(next.access_code, {
-        maxLen: 64,
-        pattern: ACCESS_CODE_PATTERN,
-      });
-      const accountCategory = next.account_category === undefined || next.account_category === null || next.account_category === ""
-        ? "organization"
-        : requireEnum(next.account_category, TENANT_ACCOUNT_CATEGORIES);
-      const planCode = next.plan_code === undefined || next.plan_code === null || next.plan_code === ""
-        ? null
-        : requireEnum(next.plan_code, TENANT_PLAN_CODES);
-      const districtSlugRaw = optionalText(next.district_slug, { maxLen: 80 });
-      const districtSlug = districtSlugRaw
-        ? requireText(normalizeDistrictSlug(districtSlugRaw), { maxLen: 63, pattern: SLUG_PATTERN })
-        : "";
-      const districtName = optionalText(next.district_name, { maxLen: 120 });
-      if ((districtSlug && !districtName) || (!districtSlug && districtName)) {
-        return jsonResponse(400, {
-          error: "District name and slug must both be provided when assigning a district.",
-        });
-      }
-      if (
-        (accountCategory === "individual" &&
-          !(!planCode || planCode === "individual_yearly" || planCode === "individual_monthly")) ||
-        (accountCategory === "district" &&
-          !(!planCode || planCode === "core" || planCode === "growth" || planCode === "enterprise")) ||
-        (accountCategory === "organization" &&
-          !(!planCode || planCode === "starter" || planCode === "scale" || planCode === "enterprise"))
-      ) {
-        return jsonResponse(400, { error: "Invalid plan for tenant account category." });
-      }
-      if (accountCategory === "individual" && (districtSlug || districtName)) {
-        return jsonResponse(400, { error: "Individual accounts cannot be assigned to a district." });
-      }
-
-      let districtId: string | null = null;
-      if (districtSlug && districtName) {
-        const districtResult = await ensureDistrict(adminClient, districtSlug, districtName);
-        if (districtResult.error) {
-          return jsonResponse(400, { error: districtResult.error });
-        }
-        districtId = districtResult.districtId;
-      }
-
-      const { data, error } = await adminClient
-        .from("tenants")
-        .update({ name, access_code: accessCode, district_id: districtId })
-        .eq("id", id)
-        .select("id, name, access_code, status, created_at, district_id, primary_admin_profile_id")
-        .single();
-
-      if (error || !data) {
-        if (isMissingDistrictIdColumn(error as PgError)) {
-          return jsonResponse(400, {
-            error: "District foundation is not enabled yet. Run the latest database migration.",
-          });
-        }
-        return jsonResponse(400, { error: "Unable to update tenant." });
-      }
-
-      const { error: policyError } = await upsertTenantPolicy(adminClient, {
-        tenant_id: id,
-        account_category: accountCategory,
-        plan_code: planCode,
-        updated_by: user.id,
-        updated_at: new Date().toISOString(),
-      });
-
-      if (policyError) {
-        console.error("super-tenant-mutate tenant policy update failed", {
-          code: policyError.code,
-          message: policyError.message,
-          details: policyError.details,
-          hint: policyError.hint,
-          accountCategory,
-          planCode,
-          tenantId: id,
-        });
-        return jsonResponse(400, {
-          error: "Unable to update tenant plan details.",
-        });
-      }
-
-      await writeAudit("update_tenant", "tenant", data.id, {
-        tenant_name: data.name,
-        account_category: accountCategory,
-        plan_code: planCode,
-        district_slug: districtSlug || null,
-      });
-
-      return jsonResponse(200, { data: (await enrichTenants([data as TenantRow]))[0] });
+    if (action === "update_tenant" || action === "set_tenant_status") {
+      return await handleTenantWriteAction(actionContext);
     }
-
-    if (action === "set_tenant_status") {
-      const next = payload;
-      const id = requireUuid(next.id);
-      const status = requireEnum(next.status, TENANT_STATUSES);
-      const superPassword = requireText(next.super_password, { maxLen: 1024 });
-      const confirmPhrase = requireText(next.confirm_phrase, { maxLen: 32 });
-
-      if (confirmPhrase !== "CONFIRM") {
-        return jsonResponse(400, { error: "Confirmation phrase mismatch." });
-      }
-
-      const verified = await verifySuperPassword(
-        supabaseUrl,
-        publishableKey,
-        profile.auth_email ?? user.email ?? "",
-        superPassword
-      );
-      if (!verified) {
-        return jsonResponse(403, { error: "Super password verification failed." });
-      }
-
-      const { data, error } = await adminClient
-        .from("tenants")
-        .update({ status })
-        .eq("id", id)
-        .select("id, name, access_code, status, created_at, district_id, primary_admin_profile_id")
-        .single();
-
-      if (error || !data) {
-        if (isMissingStatusColumn(error as PgError)) {
-          return jsonResponse(400, {
-            error:
-              "Tenant status is not enabled yet. Run the latest database migration.",
-          });
-        }
-        return jsonResponse(400, { error: "Unable to update tenant status." });
-      }
-
-      await writeAudit(
-        status === "suspended"
-          ? "suspend_tenant"
-          : status === "archived"
-            ? "archive_tenant"
-            : "reactivate_tenant",
-        "tenant",
-        data.id,
-        { tenant_name: data.name, status: data.status }
-      );
-
-      return jsonResponse(200, { data: (await enrichTenants([data as TenantRow]))[0] });
-    }
-
     if (action === "send_primary_admin_reset") {
       const next = payload;
       const tenantId = requireUuid(next.tenant_id);
